backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "job-search-agent"
version = "0.1.0"
description = "Automatisierter Jobportal-Scraper und -Evaluator für deutsche Stellenbörsen"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.28.0",
    "anthropic>=0.84.0",
    "ollama>=0.4.0",
    "beautifulsoup4>=4.12.0",
    "playwright>=1.45.0",
    "chromadb>=1.5.0",
    "duckduckgo-search>=7.0.0",
    "jinja2>=3.1.6",
    "aiofiles>=23.0.0",
    "aiosqlite>=0.20.0",
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "pytest-mock>=3.15.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.15.0",
    "mypy>=1.19.0",
    "pre-commit>=3.7.0",
    "pip-audit>=2.7.0",
    "bandit[toml]>=1.7.0",
    "types-PyYAML>=6.0.12"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = ["-v", "--tb=short"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
exclude = ["tests/"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 25
show_missing = true

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]
